# [Performance] allow for multithreaded channel extraction