Added multithreaded channel extraction giving us:

5x speedups in 8-bit data extraction (from 5s to 1s on ~1.3gigs of data)
2.2x speedups in 16-bit data extraction (from 3.5s to 1.6s on ~2gigs of data)
2.2x speedups in 32-bit data extraction (from 7s to 3.2s on ~3.7gigs of data)

Additionally across all bitdepths there should be a roughly 10% speedup in write speeds